use TSTART, TSTOP from ft1 file to set outer boundaries of Voronoi cells rather than using reflection of midpoint to neighboring event

+++ python/create_lc_file.py
@@ -5,7 +5,7 @@ used in BayesianBlocks analysis.
 @author J. Chiang
 """
 #
-# $Header: /nfs/slac/g/glast/ground/cvs/ScienceTools-scons/BayesianBlocks/python/create_lc_file.py,v 1.2 2011/09/19 03:59:35 jchiang Exp $
+# $Header: /nfs/slac/g/glast/ground/cvs/BayesianBlocks/python/create_lc_file.py,v 1.3 2011/09/19 05:07:49 jchiang Exp $
 #
 import os
 import numpy as num
@@ -17,11 +17,13 @@ gtbindef = GtApp('gtbindef')
 gtbin = GtApp('gtbin')
 gtexposure = GtApp('gtexposure')
 
-def write_cell_boundaries(arrival_times, outfile):
+def write_cell_boundaries(arrival_times, outfile, tbounds=None):
     times = num.array(arrival_times)
     cell_boundaries = ((times[:-1] + times[1:])/2.).tolist()
-    cell_boundaries.insert(0, (3*times[0] - times[1])/2.)
-    cell_boundaries.append((3*times[-1] - times[-2])/2.)
+    if tbounds is None:
+        tbounds = (3*times[0] - times[1])/2., (3*times[-1] - times[-2])/2.
+    cell_boundaries.insert(0, tbounds[0])
+    cell_boundaries.append(tbounds[1])
     output = open(outfile, 'w')
     for tmin, tmax in zip(cell_boundaries[:-1], cell_boundaries[1:]):
         output.write("%16.5f  %16.5f\n" % (tmin, tmax))
@@ -43,20 +45,25 @@ def ebounds(ft1file):
             return float(data[0]), float(data[1])
 
 def create_lc_file(evfile, scfile, lcfile, irfs, specin=-2.,
-                   tmp_ext='time_bins'):
+                   tmp_ext='time_bins', clean=True, old_tbounds=False):
     binfile = '%s.txt' % tmp_ext
     tbinfile = '%s.fits' % tmp_ext
     events = FitsNTuple(evfile)
-    write_cell_boundaries(events.TIME, binfile)
-    gtbindef.run(bintype='t', binfile=binfile, outfile=tbinfile)
     tstart, tstop = tbounds(evfile)
+    if old_tbounds:
+        write_cell_boundaries(num.sort(events.TIME), binfile)
+    else:
+        write_cell_boundaries(num.sort(events.TIME), binfile,
+                              tbounds=(tstart, tstop))
+    gtbindef.run(bintype='t', binfile=binfile, outfile=tbinfile)
     emin, emax = ebounds(evfile)
     gtbin.run(evfile=evfile, scfile=scfile, outfile=lcfile, algorithm='LC',
               tbinalg='FILE', tstart=tstart, tstop=tstop, tbinfile=tbinfile)
     gtexposure.run(infile=lcfile, scfile=scfile, irfs=irfs, srcmdl='none',
                    specin=specin, emin=emin, emax=emax)
-    os.remove(binfile)
-    os.remove(tbinfile)
+    if clean:
+        os.remove(binfile)
+        os.remove(tbinfile)
 
 if __name__ == '__main__':
     import sys
